+++ Origo.GodotAdapter.Tests/Snd/TypedDataInitializerTests.cs
@@ -1,3 +1,4 @@
+using Origo.Core.Snd.Metadata;
 using Origo.GodotAdapter.Snd;
 using Xunit;
 
@@ -6,5 +7,9 @@ namespace Origo.GodotAdapter.Tests;
 public class TypedDataInitializerTests
 {
     [Fact]
-    public void IsLoaded_ReturnsTrue() => Assert.True(TypedDataInitializer.IsLoaded);
+    public void EnsureLoaded_TriggersAdapterKindRegistration()
+    {
+        TypedDataInitializer.EnsureLoaded();
+        Assert.Equal((byte)128, TypedDataTypeMap.GetKindForType(typeof(Godot.Vector2)));
+    }
 }